requirements.txt
# 🔧 OmniDexGrasp Dependencies

# 🔬 Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# ⚙️ Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0
pyyaml>=6.0

# 🤖 3D Vision & Hand Models
# pytorch3d → manual install (see above)
manotorch @ git+https://github.com/lixiny/manotorch.git
pytorch-kinematics>=0.7.0

# 📦 3D Processing
trimesh>=4.0.0
open3d>=0.17.0
scipy>=1.10.0
scikit-learn>=1.3.0

# 🖼️ Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
pycocotools>=2.0.7

# 🌐 HTTP Client
requests>=2.31.0

# 🎨 Visualization
viser>=0.2.0

# 🍃 Optim
scikit-image>=0.20.0
imageio>=2.31.0
pyrender>=0.1.45
mesh-to-sdf>=0.0.14
chamfer-distance>=0.1
geomloss>=0.2.0
matplotlib>=3.7.0
rootutils>=1.0.0
